Earth Village: Village Center Map 1
This is the map that describes the events and activities that occur at the village center of earth village. We can start working on all the events in this part of the map so we can have a working prototype of the game to present. It has a mix of linguistic, non-linguistic, and evaluative activities.